# self optimizing routing code