--- a/conf.d/managed_endpoints.conf
+++ b/conf.d/managed_endpoints.conf
@@ -79,7 +79,7 @@ server {
 
         access_by_lua_block {
             local routing = require "routing"
-            local ds = require "lib/dataStore" 
+            local ds = require "lib/dataStore"
             routing.processCall(ds.init())
         }