--- a/scripts/lua/oauth/github.lua
+++ b/scripts/lua/oauth/github.lua
@@ -21,18 +21,17 @@
 -- A Proxy for Github OAuth API
 local cjson = require "cjson"
 local http = require "resty.http"
-local redis = require "lib/redis"
 local request = require "lib/request"
 local cjson = require "cjson"
 local utils = require "lib/utils"
-local _M = {} 
+local _M = {}
 
 function _M.process(dataStore, token)
-  local result = dataStore:getOAuthToken('github', token) 
-  if result ~= ngx.null then 
+  local result = dataStore:getOAuthToken('github', token)
+  if result ~= ngx.null then
     return cjson.decode(result)
-  end 
- 
+  end
+
   local request_options = {
     headers = {
       ["Accept"] = "application/json"
@@ -52,7 +51,7 @@ function _M.process(dataStore, token)
     request.err(500, 'OAuth provider error.')
     return
   end
- 
+
   local json_resp = cjson.decode(res.body)
   if json_resp.id == nil then
     return nil
@@ -62,10 +61,10 @@ function _M.process(dataStore, token)
     return nil
   end
 
-  dataStore:saveOAuthToken('github', token) 
+  dataStore:saveOAuthToken('github', token, cjson.encode(json_resp))
   -- convert Github's response
   -- Read more about the fields at: https://developers.google.com/identity/protocols/OpenIDConnect#obtainuserinfo
   return json_resp
 end
 
-return _M 
+return _M