--- a/scripts/lua/policies/rateLimit.lua
+++ b/scripts/lua/policies/rateLimit.lua
@@ -26,7 +26,7 @@ local request = require "lib/request"
 local _M = {}
 
 --- Limit a resource/api/tenant, based on the passed in rateLimit object
--- @param red redis client instance
+-- @param dataStore the datastore object
 -- @param obj rateLimit object containing interval, rate, scope, subscription fields
 -- @param apiKey optional api key to use if subscription is set to true
 function limit(dataStore, obj, apiKey)