--- a/scripts/lua/policies/security.lua
+++ b/scripts/lua/policies/security.lua
@@ -27,6 +27,7 @@ local _M = {}
 local request = require "lib/request"
 local utils = require "lib/utils"
 --- Allow or block a request by calling a loaded security policy
+-- @param dataStore the datastore object
 -- @param securityObj an object out of the security array in a given tenant / api / resource
 function process(dataStore, securityObj)
   local ok, result = pcall(require, utils.concatStrings({'policies/security/', securityObj.type}))