Optimizes call to routing with a prefix tree rather than just getting all products in a tenant and iterating over them trying to match the pattern.

I called this on one of our tenants with a bunch of apis: 

Optimizations off:
```
Running 5s test @ http://9.37.200.19:32801/api/a27b9526a7c9d44fd1f2550d70651d7509ed3e43b3add35af69f20ea46f5c8d9/helloWorld/36717
  2 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.23s   378.84ms   4.97s    62.71%
    Req/Sec    38.62     22.02    80.00     69.23%
  59 requests in 5.05s, 62.57KB read
Requests/sec:     11.69
Transfer/sec:     12.40KB
```

Optimizations On: 

```
Running 5s test @ http://9.37.200.19:32803/api/a27b9526a7c9d44fd1f2550d70651d7509ed3e43b3add35af69f20ea46f5c8d9/helloWorld/36717
  2 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.68s     1.14s    4.03s    51.22%
    Req/Sec   354.18    312.17     1.22k    69.23%
  2384 requests in 5.05s, 2.47MB read
Requests/sec:    472.09
Transfer/sec:    500.63KB
```