@taylorking 2017-05-01
Got a weird flame graph out of this one.. 
![out](https://cloud.githubusercontent.com/assets/1733912/25584730/344e871a-2e65-11e7-9424-e9b9cc6fde4d.png)