@taylorking 2017-05-18
under optimal conditions this is really really fast 

```
  8 threads and 10000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   121.20ms  150.71ms 626.33ms   68.84%
    Req/Sec     1.40k     1.69k    7.23k    86.29%
  82390 requests in 10.06s, 20.74MB read
  Socket errors: connect 8987, read 0, write 0, timeout 0
Requests/sec:   8190.09
Transfer/sec:      2.06MB
```