@mhamann 2017-05-28
@taylorking could you rebase this? We'll see about merging it this coming week.