@taylorking 2017-05-30
Let's wait a bit before merging this, there are a couple of our tests failing I think because of cache invalidation timing, want to get those passing before we merge this