@taylorking 2017-05-30
Tests are looking good now, I want to clean up the code so that it preserves alex's stuff for cloud foundry apps