@mhamann 2017-05-31
@taylorking looks like there are a couple of failing unit tests

```
not ok 19 - Testing routing module should find the correct redis key
# scripts/lua/routing.lua @ 45
# Failure message: ...penwhisk-apigateway/tests/../scripts/lua/routing.lua:115: attempt to compare number with nil
not ok 20 - Testing routing module should return nil if redis key doesn't exist
# scripts/lua/routing.lua @ 56
# Failure message: ...penwhisk-apigateway/tests/../scripts/lua/routing.lua:115: attempt to compare number with nil
not ok 21 - Testing routing module should find correct key when basePath is "/"
# scripts/lua/routing.lua @ 64
# Failure message: ...penwhisk-apigateway/tests/../scripts/lua/routing.lua:115: attempt to compare number with nil
not ok 22 - Testing routing module should find correct key when resourcePath is "/"
# scripts/lua/routing.lua @ 75
# Failure message: ...penwhisk-apigateway/tests/../scripts/lua/routing.lua:115: attempt to compare number with nil
```