@mhamann 2017-06-10
Seems like this has stabilized. @taylorking can you rebase? Do you feel this is good to go?